requirements.txt
requests
slack-sdk
notion-client
google-auth
google-api-python-client
